Reto12/requirements.txt
django
djangorestframework
psycopg2-binary
django-rest-auth
django-allauth
djangorestframework-simplejwt